requirements.txt
flask==3.1.3
jinja2==3.1.6
playwright==1.58.0
deep-translator==1.11.4
Pillow==11.0.0
gunicorn==21.2.0
